Log to Databases via DBI

Log to a database table with any DBI compatabile backend.

Usage

x <- AppenderDbi$new(conn, table, threshold = NA_integer_,
  layout = select_dbi_layout(conn), close_on_exit = TRUE)
    x$set_close_on_exit(x)
x$show(n = 20, threshold = NA_integer_)
x$append(event)
x$clone(deep = FALSE)
    x$close_on_exit
x$conn
x$data
x$destination
x$table

Creating a New Appender

conn

a DBI connection

threshold

character or integer scalar. The minimum log level that triggers this logger. See log_levels

layout

a Layout that will be used for formatting the LogEvents passed to this Appender

Fields and Methods

close_on_exit, set_close_on_exit()

TRUE or FALSE. Close the Database connection when the Logger is removed?

data

Querry the whole log from the Database and return it as a data.frame

conn

get the DBI connection object

table

Name of the target database table

show(n, threshold)

Show the last n log entries with a log level bellow threshold. The log entries will be formated for console output via the defined Layout

data

Get the log recorded by this Appender as a data.table with a maximum of buffer_size rows

See also