Log to an In-Memory Data.Table

An Appender that outputs to an in-memory data.table. This requires that you have the suggested package data.table installed. This kind of Appender is useful for interactive use, and has very little overhead.

Custom Fields

AppenderDt supports custom fields, but they have to be pre-allocated in the prototype argument. Custom fields that are not part of the prototype are discarded.

With the default settings, the custom field value is included in the data.table as a list column to store arbitrary R objects (see example). It is recommended to use this feature only TRACE level.

Usage

x <- AppenderDt$new(threshold = NA_integer_, layout = LayoutFormat$new(fmt =
  "%L [%t] %m", timestamp_fmt = "%H:%M:%S", colors = getOption("yog.colors")),
  prototype = data.table::data.table(.id = NA_integer_, level = NA_integer_,
  timestamp = Sys.time(), caller = NA_character_, msg = NA_character_, value =
  list(list())), buffer_size = 1e+05)
    x$append(event)
x$show(n = 20, threshold = NA_integer_)
x$clone(deep = FALSE)
    x$data
x$destination

Creating a New Appender

buffer_size

integer scalar. Number of rows of the in-memory data.table

prototype

A prototype data.table. This is only necessary to set manually if you use custom LogEvents. The default prototype already contains a list column called value that can hold arbitrary R values (see examples)

threshold

character or integer scalar. The minimum log level that triggers this logger. See log_levels

layout

a Layout that will be used for formatting the LogEvents passed to this Appender

Fields and Methods

show(n, threshold)

Show the last n log entries with a log level bellow threshold. The log entries will be formated for console output via the defined Layout

data

Get the log recorded by this Appender as a data.table with a maximum of buffer_size rows

See also

Examples

lg <- Logger$new(
  "test",
  appenders = list(memory = AppenderDt$new()),
  threshold = NA,
  parent = NULL  # to prevent routing to root logger for this example
)
lg$debug("test")
lg$error("test")

# Displaying the log
lg$appenders$memory$data

#>    .id level           timestamp caller  msg value
#> 1:   1   500 2018-12-26 01:06:01   eval test    NA
#> 2:   2   200 2018-12-26 01:06:01   eval test    NA

lg$appenders$memory$show()

#> DEBUG [01:06:01] test
#> ERROR [01:06:01] test

show_log(target = lg$appenders$memory)

#> DEBUG [01:06:01] test
#> ERROR [01:06:01] test

# If you pass a Logger to show_log, it looks for the first MemoryAppender it
# can find.
show_log(target = lg)

#> DEBUG [01:06:01] test
#> ERROR [01:06:01] test

# Data tables can store arbitrary R values in list columns. The default
# AppenderDt comes with a single list column called "value" that you can use
lg$info("the iris data frame", value = iris)
lg$appenders$memory$data

#>    .id level           timestamp caller                 msg        value
#> 1:   1   500 2018-12-26 01:06:01   eval                test           NA
#> 2:   2   200 2018-12-26 01:06:01   eval                test           NA
#> 3:   3   400 2018-12-26 01:06:01   eval the iris data frame <data.frame>

head(lg$appenders$memory$data$value[[3]])

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa