Log to Databases via RJDBC

Log to a database table with the RJDBC package. RJDBC is only somewhat DBI compliant and does not work with AppenderDbi. I personally do not recommend using RJDBC if it can be avoided.

Usage

x <- AppenderRjdbc$new(conn, table, threshold = NA_integer_,
  layout = select_dbi_layout(conn), close_on_exit = TRUE)
    x$append(event)
x$clone(deep = FALSE)
    x$

Creating a New Appender

conn

an RJDBC connection

threshold

character or integer scalar. The minimum log level that triggers this logger. See log_levels

layout

a Layout that will be used for formatting the LogEvents passed to this Appender

Fields and Methods

close_on_exit, set_close_on_exit()

...

data

...

conn

get the DBI connection object

table

Name of the target database table

show(n, threshold)

Show the last n log entries with a log level bellow threshold. The log entries will be formated for console output via the defined Layout

data

Get the log recorded by this Appender as a data.table with a maximum of buffer_size rows

See also