Format Log Events for Output to Databases

Format a LogEvent as data.frame for inserting into a Database.

Usage

x <- LayoutDbi$new(event_vals = c("level", "timestamp", "caller", "msg"),
  logger_vals = character(), col_types = NULL)
    x$format_event(event)
x$set_col_types(x)
x$sql_create_table(table)
x$clone(deep = FALSE)
    x$col_names
x$col_types

Creating a New Layout

col_types

a named character vector of column types supported by the target database. Must include all columns described in event_vals, logger_vals. If a new database table is created by LayoutDbi its column order will correspond to the order of col_types.

event_vals

Names of the fields of the LogEvent to include in the output object.

logger_vals

Names of the fields of the Logger that produced the LogEvent to include in the output object.

Fields and Methods

col_types, set_col_types()

Get/set the col_types of this Layout

col_names

Convenience method to get the names of the col_types vector

format_event(event)

format a LogEvent

Database Specific Layouts

Different databases have different data types and features. Currently the following LayoutDBI subclasses exist that deal with specific databases, but this list is expected to grow as yog matures:

LayoutSQLite: Needs its own Layout because SQLite does not support timestamps

LayoutDBI: For all other datbases

The utility function select_dbi_layout() returns the appropriate Layout for a DBI connection.

See also

Examples

# setup a dummy LogEvent
event <- LogEvent$new(
  logger = Logger$new("dummy logger", user = "testuser"),
  level = 200,
  timestamp = Sys.time(),
  caller = NA_character_,
  msg = "a test message"
)

# defaults
lo <- LayoutDbi$new()
lo$format_event(event)

#>   level           timestamp caller            msg
#> 1   200 2018-12-24 09:44:35   <NA> a test message

# SQLite does not support timestamps so LayoutSqlite converts them to text
lo <- LayoutSqlite$new()
str(lo$format_event(event))

#> 'data.frame':	1 obs. of  4 variables:
#>  $ level    : num 200
#>  $ timestamp: chr "2018-12-24 09:44:35"
#>  $ caller   : chr NA
#>  $ msg      : chr "a test message"

# advanced example that supports a custom_field:
lo <- LayoutDbi$new(
  event_vals = c("level", "timestamp", "msg", "custom_field"),
  logger_vals = c(logger_name = "name", "user"),
  col_types =  c(
    timestamp = "timestamp",
    user = "varchar(256)",
    logger_name = "varchar(256)",
    level = "smallint",
    msg = "varchar(2048)",
    custom_field = "integer"
  )
)

event <- LogEvent$new(
  logger = Logger$new("dummy logger", user = "testuser"),
  level = 200,
  timestamp = Sys.time(),
  caller = NA_character_,
  msg = "a test message",
  custom_field = "blubb"
)

lo$format_event(event)

#>             timestamp     user  logger_name level            msg custom_field
#> 1 2018-12-24 09:44:35 testuser dummy logger   200 a test message        blubb

#'