Loggers

A Logger records the log message and some metadata (timestamp, calling function) as a LogEvent and passes this event on to one or several Appenders that write the event to a destination (a file, the console, ...). All Loggers of an R session are in a hierarchical structure, and each Logger (except the Root Looger) passes on LogEvents to the Appenders of it ancestral Loggers. See vignette("yog", package = "yog") for more info.

Usage

x <- Logger$new(name, appenders = list(), threshold = 400L, user = get_user(),
  parent = yog::yog, exception_handler = default_exception_handler,
  propagate = TRUE)
    x$log(level, msg, ..., timestamp = Sys.time(), caller = get_caller(-7))
x$fatal(msg, ...)
x$error(msg, ...)
x$warn(msg, ...)
x$info(msg, ...)
x$debug(msg, ...)
x$trace(msg, ...)
x$add_appender(appender, name = NULL)
x$remove_appender(pos)
x$handle_exception(...)
x$set_name(x)
x$set_exception_handler(fun)
x$set_propagate(x)
x$set_parent(logger)
x$set_threshold(level)
x$set_appenders(x)
x$set_user(x)
    x$ancestry
x$appenders
x$exception_handler
x$inherited_appenders
x$last_event
x$name
x$parent
x$propagate
x$threshold
x$user

Creating Loggers

If you want logging for a Project (f.e a Package you are developing) that is separate from the global logging, you can create a new logger with Logger$new(). If you just want to add different outputs (for example logfiles) to the Root Logger, look into Appenders.

Fields

You can either specify the fields in Logger$new() or modify them after creation with setter functions of the form logger$set_fieldname(value) (see examples)

name

character scalar. Name of the Logger. Should be unique amongst Loggers. If you define a logger for an R Package, the logger should have the same name as the Package.

appenders

list of Appenders. The appenders used by this logger to write log entries to the console, to files, etc...

threshold

character or integer scalar. The minimum log level that triggers this logger

user

character scalar. The current user name or email adress. This information can be used by the appenders

parent

a Logger. Usually the Root logger. All Loggers must be descentents of the Root logger for yog to work as intended.

exception_handler

a function that takes a single argument e. The function used to handle errors that occur durring loging. Defaults to demoting errors to warnings

propagate

TRUE or FALSE. Should log messages be passed on to the appenders of the ancestral Loggers?

appenders

A list of all Appenders attached to the current logger

ancestry

A named logical vector of containing the propagate value of each Logger upper the inheritance tree. The names are the names of the appenders.

inherited_appenders

A list of all inherited appenders from ancestral Loggers of the current Logger

filters

a list of predicates (functions that return either TRUE or FALSE). If all of these functions evaluate to TRUE the LogEvent is passed on to the Logger's Appenders. Since LogEvents have reference semantics, filters can also be abused to modify LogEvents before they are passed on. Look at the source code of with_log_level() or with_log_value() for examples.

last_event

The last LogEvent produced by the current Logger

name

Name of the Logger. Mainly used for display purposes. If you define a Logger for a package, this should be the same name as the packages

parent

Parent Logger of the current Logger (NULL for the Root Logger)

threshold

An integer. Threshold of the current Logger (i.e the maximum log level that this Logger processes)

user

The current user

Methods

fatal(msg, ...)

Logs a message with level fatal on this logger. If there are unnamed arguments in ..., they will be pased to base::sprintf() along with message. Named arguments will be passed as custom fields to LogEvent. If there are named arguments the names must be unique

error(msg, ...)

Logs a message with level error on this logger. The arguments are interpreted as for fatal().

warn(msg, ...)

Logs a message with level warn on this logger. The arguments are interpreted as for fatal().

info(msg, ...)

Logs a message with level info on this logger. The arguments are interpreted as for fatal().

debug(msg, ...)

Logs a message with level debug on this logger. The arguments are interpreted as for fatal().

trace(msg, ...)

Logs a message with level trace on this logger. The arguments are interpreted as for fatal().

log(level, msg, ..., timestamp, caller)

If the level passes the Logger threshold a new LogEvent with level, msg, timestamp and caller is created. Unnamed arguments in ... will be combined with msg via base::sprintf(). Named arguments in ... will be passed on to LogEvent$new() as custom fields.
If the new LogEvent passes this Loggers filters, it will be dispatched
to the relevant [Appenders] and checked against their thresholds and
filters.

add_appender(appender, name = NULL)

Adds a new Appender to the Logger. appender must be an Appender object. name is optional and will be used as name in the list of appenders, i.e if you do logger$add_appender(AppenderConsole$new(), name = "console") you can refer to it via logger$appenders$console.

remove_appender(pos)

Removes and Appender from a Logger. pos can be an integer or character vector referring either to the positions or names of the Appenders to be removed.

exception_handler(e)

Used to handle errors that occur durring the logging process. The defaul is to convert errors to warnings.

set_name(x)

Set the Logger name to the character scalar x

filter(x)

Determine whether the LogEvent x should be passed on to Appenders (TRUE) or not (FALSE). See also the active binding filters

set_filters(filters)

set a list of filters. The filters must be functions that take exactly two named arguments: the LogEvent event and the Logger obj.

Examples

yog$info("Today is %s", Sys.Date() )

#> INFO  [10:13:26.220] Today is 2018-12-23

# yog includes a pre-configured root logger
yog$fatal("This is a serious error")

#> FATAL [10:13:26.226] This is a serious error

# if you want to take advantage of hierarchical logging, you can create new loggers.
#' # the following creates a new logger that logs to a temporary file.
tf <- tempfile()
mylogger <- Logger$new(
  "mylogger",
  appenders = AppenderFile$new(tf)
)

# The new logger passes the log message on to the appenders of its parent
# logger, which is by default the root logger. This is why the following
# writes not only the file 'tf', but also to the console.
mylogger$fatal("blubb")

#> FATAL [10:13:26.248] blubb

readLines(tf)

#> [1] "FATAL [2018-12-23 10:13:26.248] blubb"

# use string inerpolation and custom fields
tf2 <- tempfile()
mylogger$add_appender(AppenderFile$new(tf2, layout = LayoutJson$new()))
mylogger$info("Not all %s support custom fields", "appenders", type = "test")

#> INFO  [10:13:26.264] Not all appenders support custom fields

readLines(tf)

#> [1] "FATAL [2018-12-23 10:13:26.248] blubb"                                  
#> [2] "INFO  [2018-12-23 10:13:26.264] Not all appenders support custom fields"

readLines(tf2)

#> [1] "{\"level\":400,\"timestamp\":\"2018-12-23 10:13:26\",\"caller\":\"eval\",\"msg\":\"Not all appenders support custom fields\",\"type\":\"test\"}"