Simple Logging

These functions provide a simple interface to the root logger. If you do not need any of the more advanced features of yog, start here.

Yog provides convenience functions to manage the root Logger. These are intended for interactive use, and for people who just need basic logging facilities and don't want to worry about hierarchical loggers and R6 classes.

threshold() sets or retrieves the threshold for an Appender or Logger (the minimum level of log messages it processes). It's target defaults to the root logger.

console_threshold() is a shortcut to set the threshold of the root loggers AppenderConsole, which is usually the only Appender that manages console output for a given R session.

add_appender() and remove_appender() add Appenders to Loggers and other Appenders.

FATAL(msg, ...)

ERROR(msg, ...)

WARN(msg, ...)

INFO(msg, ...)

DEBUG(msg, ...)

TRACE(msg, ...)

log_exception(...)

threshold(level, target = yog::yog)

console_threshold(level, target = yog::yog$appenders$console)

add_appender(appender, name = NULL, target = yog::yog)

remove_appender(pos, target = yog::yog)

show_log(n = 20, threshold = NA, target = yog::yog)

Arguments

msg, ...
passed on to base::sprintf()

level, threshold
an integer or character scalar, see getOption("yog.log_levels") for possible values. For threshold 0 ("off") and NA ("all") are also valid.

target
a Logger or Appender. Defaults to the root logger.

appender
an Appender

name
character scalar. An optional name for the new Appender.

pos
integer index or character names of the appenders to remove

n
integer scalar. Show only the last n log entries that match threshold

Value

FATAL() ... TRACE() and log_exception() return the log message as a character vector.

threshold() and console_threshold() return the log_level of target as integer (invisibly)

add_appender() and remove_appender() return target.

Examples

FATAL("This is an important message about %s going wrong", "something")

#> FATAL [18:45:15.323] This is an important message about something going wrong

DEBUG("Debug messages are hidden by default")
console_threshold("debug")  # you must use lower case names here
DEBUG("Unless we lower the threshold")

#> DEBUG [18:45:15.327] Unless we lower the threshold

add_appender(AppenderConsole$new(), "second_console_appender")
FATAL("Multiple console appenders are a bad idea")

#> FATAL [18:45:15.332] Multiple console appenders are a bad idea
#> FATAL [18:45:15.332] Multiple console appenders are a bad idea

remove_appender("second_console_appender")
INFO("Good that we defined an appender name, so it's easy to remove")

#> INFO  [18:45:15.341] Good that we defined an appender name, so it's easy to remove