Events - The Atomic Unit of Logging

A LogEvent is a single unit of data that should be logged. It usually contains at minimum the log_level, a timestamp, the name of the calling function, the log message and a reference to the Logger that created it. LogEvents are created by a Logger, and then processed by an Appenders.

Usage

x <- LogEvent$new(logger, level = 400, timestamp = Sys.time(), caller = NA,
  msg = NA, ...)
    x$clone(deep = FALSE)
    x$caller
x$level
x$level_name
x$logger
x$msg
x$timestamp
x$values

Creating LogEvents

not yet stable Currently custom LogEvents are not officially supported in yog, but a lot of the infrastructure to make them possible is already in place. If you require this function, please upvote this issue

logger

Logger scalar. The Logger this LogEvent is associated with

Fields

Log Evnets contain the following data fields

level: The log_level of the event

timestamp: POSIXct.

caller: character. the function that called this event

msg: character. the log message.

logger the Logger that created the LogEvent. See examples.

Usually the above values will be scalars, but (except for "logger") they can also be vectors if they are all of the same length (or scalars that will be recycled). In this case the event will be treated by the Appenders and Layouts as if several separate events.

Log events also contain the following active bindings:

value: an active binding that returns a named list containing all the above values.

level_name the log level as a character string.

Examples

# Usually you do not create a LogEvent in this manner, it is rather provided
# by the Logger

l <- Logger$new("dummy logger", appenders = NULL)
l$error("foo bar")

#> ERROR [09:44:38.686] foo bar

# The last LogEvent produced by a logger is saved to the last_event field
l$last_event

#> ERROR [2018-12-24 09:44:38] foo bar

l$last_event$level

#> [1] 200

l$last_event$level_name

#>     200 
#> "error" 

l$last_event$msg

#> [1] "foo bar"

# Also contains a reference to the complete logger that created it
l$last_event$logger$name

#> [1] "dummy logger"

l$last_event$logger$user

#> [1] "stefan.b.fleck@gmail.com"

# This is really a reference to the complete logger, so the following is
# also possible (though nonsensical)
l$last_event$logger$last_event$msg

#> [1] "foo bar"

identical(l, l$last_event$logger)

#> [1] TRUE

# Because LogEvents can also be vectors, the following works:
l$warn(c("a", "vector", "log", "message"))

#> WARN  [09:44:38.705] a
#> WARN  [09:44:38.705] vector
#> WARN  [09:44:38.705] log
#> WARN  [09:44:38.705] message