A Fully Featured Logging Framework for R

A flexible R6 based logging framework. Yog is under heavy development right now, please don't heistate to contact the maintainer with bug reports, wishes, suggestions, etc...

Options

You can also set these options in your .Rprofile to make them permanent

yog.colors

a list of functions used for coloring the log levels in console output. Usually these will be functions from the package crayon

yog.log_levels

A named integer vector of log levels that are known to yog for labeling, setting thresholds, etc... . Instead of modifying this option manually use add_log_levels() and remove_log_levels()

yog.suspend_logging

TRUE or FALSE. Suspend all logging for all loggers. Instead of modifying this option manually use suspend_logging() and unsuspend_logging()

yog.user

a character scalar. The default username for new Loggers. Users can also be set per Logger, though this rarely makes sense.

See also